removed InitOptions.ts
