+++ src/commands/ChatCommand.ts
@@ -1,10 +1,9 @@
 import { Interaction } from "discord.js";
-import { ChatCommandInit } from "./types/InitOptions";
-import { DefaultParameter, Parameter } from "../structures/Parameter";
+import { DefaultParameter, Parameter, ParameterSchema } from "../structures/Parameter";
 import { ChatCommandObject, TextCommandOptionChoiceObject, ChatCommandOptionObject, ChatCommandOptionType } from "../structures/types/api";
 import { CommandRegExps } from "./types/commands";
 import { CommandManager } from "../structures/CommandManager";
-import { PermissionGuildCommand } from "./base/PermissionGuildCommand";
+import { PermissionGuildCommand, PermissionGuildCommandInit } from "./base/PermissionGuildCommand";
 import { InputManager } from "../structures/InputManager";
 import { DEFAULT_BLANK_DESCRIPTION } from "../constants";
 import { CommandGroup } from "./CommandGroup";
@@ -335,4 +334,42 @@ export class ChatCommand extends PermissionGuildCommand {
             });
         return usageTemplate;
     }
+}
+
+/**
+ * Intialization options of chat command
+ * @interface
+ * @extends {PermissionGuildCommandInit}
+ */
+export interface ChatCommandInit extends PermissionGuildCommandInit {
+    /**
+     * List of object defining all parameters of the command
+     * @type {?ParameterSchema[] | "simple" | "no_input"}
+     */
+    parameters?: ParameterSchema[] | "simple" | "no_input";
+    /**
+     * Different string that can be used with prefix to invoke the command
+     * @type {?Array<string>}
+     */
+    aliases?: string[] | string;
+    /**
+     * Command description
+     * @type {?string}
+     */
+    description?: string;
+    /**
+     * Command usage (if *undefined*, the usage will be automatically generated using parameters)
+     * @type {?string}
+     */
+    usage?: string;
+    /**
+     * Whether this command is visible in the help message
+     * @type {?boolean}
+     */
+    visible?: boolean;
+    /**
+     * Whether this command should be registered as a slash command
+     * @type {?boolean}
+     */
+    slash?: boolean;
 }
